requirements.txt
# LangGraph Customer Support Agent Dependencies

langgraph>=0.0.60
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# MCP (Model Context Protocol) Support
mcp[cli]
langchain-mcp-adapters>=0.1.9

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
typing-extensions>=4.8.0
datetime
uuid
jsonschema>=4.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# API and HTTP
httpx>=0.25.0
aiohttp>=3.8.0

# Logging and Monitoring
structlog>=23.0.0
rich>=13.0.0
